fixed model updating stuff and added a todo

+++ backend/src/main/kotlin/space/davids_digital/cloud_computing_lab/backend/service/AwsLambdaModelControlService.kt
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service
 @Service
 @Qualifier("aws-lambda")
 class AwsLambdaModelControlService: ModelControlService {
-    override fun applyDataset(agentId: Int, dataId: Int) {
+    override fun applyDataset(agentId: Int, dataId: Long) {
         TODO("Not yet implemented")
     }
 }
