+++ backend/src/main/kotlin/space/davids_digital/cloud_computing_lab/backend/service/ModelControlService.kt
@@ -1,5 +1,5 @@
 package space.davids_digital.cloud_computing_lab.backend.service
 
 interface ModelControlService {
-    fun applyDataset(agentId: Int, dataId: Int)
+    fun applyDataset(agentId: Int, dataId: Long)
 }
